requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
airportsdata==20250706
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
astor==0.8.1
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
blake3==1.0.5
blessed==1.21.0
blobfile==3.0.0
build==1.2.2.post1
cachetools==6.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
compressed-tensors==0.10.2
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
decorator==5.2.1
decord==0.6.0
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.0
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.2
fastrlock==0.8.3
filelock==3.18.0
fonttools==4.58.5
frozenlist==1.7.0
fsspec==2024.6.1
gguf==0.17.1
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.5
hf_transfer==0.1.9
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
importlib_metadata==8.7.0
interegular==0.3.3
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lark==1.2.2
litellm==1.74.0.post1
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.10.11
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistral_common==1.7.0
modelscope==1.28.0
mpmath==1.3.0
msgpack==1.1.1
msgspec==0.19.0
multidict==6.6.3
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.3
ninja==1.11.1.4
numba==0.61.2
numpy==2.1.2
openai==1.90.0
opencv-python-headless==4.12.0.88
orjson==3.10.18
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.3.1
parso==0.8.4
partial-json-parser==0.2.1.1.post6
peft==0.16.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.8
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pybase64==1.4.1
pycountry==24.6.1
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pynvml==12.0.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.0
ray==2.47.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.5.1
rpds-py==0.26.0
safetensors==0.5.3
scipy==1.15.3
sentencepiece==0.2.0
sentry-sdk==2.32.0
setproctitle==1.3.6
# sgl-kernel==0.2.4
# sglang==0.4.9.post1
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.13.1
stack-data==0.6.3
starlette==0.46.2
sympy==1.13.3
tiktoken==0.9.0
timm==1.0.16
tokenizers==0.21.2
tomli==2.2.1
traitlets==5.14.3
trl==0.19.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
uv==0.7.20
uvicorn==0.35.0
uvloop==0.21.0
wandb==0.21.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
# xformers==0.0.30
# xgrammar==0.1.19
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
easydict==1.13
evaluate==0.4.5
scikit-learn==1.7.1
# liger-kernel==0.6.0
